+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.6.6 - Unreleased
 
+### Added
+* `kwimage.draw_text_on_image` now support `halign`
+
 
 ## Version 0.6.5 - Released 2020-08-26